Изменил структуру проекта
